+++ ui/mods/com.pa.quitch.gwaioverhaul/shared/ai_inventory.js
@@ -3,12 +3,12 @@ define([
   "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/unit_groups.js",
 ], function (gwoUnit, gwoGroup) {
   const commanderUnits = [gwoUnit.commander];
-  const legonisUnitsNotMobile = [
+  const legonisUnitsImmobile = [
     gwoUnit.vehicleFactory,
     gwoUnit.vehicleFactoryAdvanced,
   ];
   const legonisUnitsMobile = gwoGroup.vehiclesMobile.concat(gwoUnit.ares);
-  const foundationUnitsNotMobile = [
+  const foundationUnitsImmobile = [
     gwoUnit.airFactory,
     gwoUnit.airFactoryAdvanced,
     gwoUnit.navalFactory,
@@ -18,17 +18,18 @@ define([
     gwoUnit.zeus,
     gwoUnit.squall
   );
-  const foundationUnitsMobileNotAir = _.without(
+  const foundationUnitsMobileNaval = _.without(
     gwoGroup.navalMobile,
     gwoUnit.squall
   );
-  const synchronousUnitsNotMobile = [
+  const synchronousUnitsImmobile = [
     gwoUnit.botFactoryAdvanced,
     gwoUnit.botFactory,
   ];
   const synchronousUnitsMobile = gwoGroup.botsMobile.concat(gwoUnit.atlas);
-  const revenantsUnitsNotMobileNoAmmo = [
+  const revenantsUnitsImmobile = [
     gwoUnit.anchor,
+    gwoUnit.antiNukeLauncher,
     gwoUnit.catalyst,
     gwoUnit.deepSpaceOrbitalRadar,
     gwoUnit.halley,
@@ -39,26 +40,24 @@ define([
     gwoUnit.umbrella,
     gwoUnit.unitCannon,
   ];
-  const revenantsUnitsNotMobileWithAmmo = [
-    gwoUnit.anchor,
-    gwoUnit.catalyst,
-    gwoUnit.deepSpaceOrbitalRadar,
-    gwoUnit.halley,
-    gwoUnit.jig,
+  const revenantsUnitsImmobileWithAmmo = revenantsUnitsImmobile.concat(
+    gwoUnit.antiNukeLauncherAmmo,
+    gwoUnit.nukeLauncherAmmo
+  );
+  const revenantsUnitsFactories = [
+    gwoUnit.antiNukeLauncher,
     gwoUnit.nukeLauncher,
-    gwoUnit.nukeLauncherAmmo,
     gwoUnit.orbitalFactory,
     gwoUnit.orbitalLauncher,
-    gwoUnit.umbrella,
-    gwoUnit.unitCannon,
   ];
   const revenantsUnitsMobile = gwoGroup.orbitalMobile.concat(gwoUnit.helios);
-  const clusterCommanders = commanderUnits.concat([
+  const clusterCommanders = commanderUnits.concat(
     gwoUnit.angel,
-    gwoUnit.colonel,
-  ]);
+    gwoUnit.colonel
+  );
   const clusterUnitsMobile = clusterCommanders;
-  const clusterUnitsNotMobile = gwoGroup.structures.concat(gwoUnit.ragnarok);
+  const clusterUnitsImmobile = gwoGroup.structures.concat(gwoUnit.ragnarok);
+  const clusterUnitsFactories = gwoGroup.factories;
 
   const commanderAmmo = [
     gwoUnit.commanderAAAmmo,
@@ -133,6 +132,7 @@ define([
     gwoUnit.stitchBuildArm,
   ];
   const revenantsBuildArms = [
+    gwoUnit.antiNukeLauncherBuildArm,
     gwoUnit.nukeLauncherBuildArm,
     gwoUnit.orbitalFabberBuildArm,
     gwoUnit.orbitalFactoryBuildArm,
@@ -160,38 +160,39 @@ define([
     commanderWeapons: commanderWeapons,
     commanderBuildArms: commanderBuildArms,
 
-    legonisUnitsNotMobile: legonisUnitsNotMobile,
+    legonisUnitsImmobile: legonisUnitsImmobile,
     legonisUnitsMobile: legonisUnitsMobile,
-    legonisUnits: legonisUnitsNotMobile.concat(legonisUnitsMobile),
+    legonisUnits: legonisUnitsImmobile.concat(legonisUnitsMobile),
     legonisWeapons: commanderWeapons,
     legonisAmmo: legonisAmmo.concat(commanderAmmo),
     legonisBuildArms: legonisBuildArms.concat(commanderBuildArms),
 
-    foundationUnitsNotMobile: foundationUnitsNotMobile,
+    foundationUnitsImmobile: foundationUnitsImmobile,
     foundationUnitsMobileAir: foundationUnitsMobileAir,
-    foundationUnitsMobileNotAir: foundationUnitsMobileNotAir,
+    foundationUnitsMobileNaval: foundationUnitsMobileNaval,
     foundationUnits: foundationUnitsMobileAir.concat(
-      foundationUnitsMobileNotAir,
-      foundationUnitsNotMobile
+      foundationUnitsMobileNaval,
+      foundationUnitsImmobile
     ),
     foundationWeapons: foundationWeapons.concat(commanderWeapons),
     foundationAmmo: foundationAmmo.concat(commanderAmmo),
     foundationBuildArms: foundationBuildArms.concat(commanderBuildArms),
 
-    synchronousUnitsNotMobile: synchronousUnitsNotMobile,
+    synchronousUnitsImmobile: synchronousUnitsImmobile,
     synchronousUnitsMobile: synchronousUnitsMobile,
-    synchronousUnits: synchronousUnitsMobile.concat(synchronousUnitsNotMobile),
+    synchronousUnits: synchronousUnitsMobile.concat(synchronousUnitsImmobile),
     synchronousWeapons: synchronousWeapons.concat(commanderWeapons),
     synchronousAmmo: synchronousAmmo.concat(commanderAmmo),
     synchronousBuildArms: synchronousBuildArms.concat(commanderBuildArms),
 
-    revenantsUnitsNotMobileNoAmmo: revenantsUnitsNotMobileNoAmmo,
-    revenantsUnitsNotMobileWithAmmo: revenantsUnitsNotMobileWithAmmo,
+    revenantsUnitsImmobile: revenantsUnitsImmobile,
+    revenantsUnitsImmobileWithAmmo: revenantsUnitsImmobileWithAmmo,
     revenantsUnitsMobile: revenantsUnitsMobile,
     revenantsUnitsWithAmmo: revenantsUnitsMobile.concat(
-      revenantsUnitsNotMobileWithAmmo
+      revenantsUnitsImmobileWithAmmo
     ),
-    revenantsUnits: revenantsUnitsMobile.concat(revenantsUnitsNotMobileNoAmmo),
+    revenantsUnitsFactories: revenantsUnitsFactories,
+    revenantsUnits: revenantsUnitsMobile.concat(revenantsUnitsImmobile),
     revenantsWeapons: revenantsWeapons.concat(commanderWeapons),
     revenantsAmmo: revenantsAmmo.concat(commanderAmmo),
     revenantsBuildArms: revenantsBuildArms.concat(commanderBuildArms),
@@ -199,8 +200,9 @@ define([
     clusterCommanders: clusterCommanders,
     clusterCommanderAmmo: clusterCommanderAmmo,
     clusterUnitsMobile: clusterUnitsMobile,
-    clusterUnitsNotMobile: clusterUnitsNotMobile,
-    clusterUnits: clusterUnitsMobile.concat(clusterUnitsNotMobile),
+    clusterUnitsImmobile: clusterUnitsImmobile,
+    clusterUnitsFactories: clusterUnitsFactories,
+    clusterUnits: clusterUnitsMobile.concat(clusterUnitsImmobile),
     clusterWeapons: clusterWeapons.concat(commanderWeapons),
     clusterAmmo: clusterAmmo.concat(commanderAmmo, clusterCommanderAmmo),
     clusterBuildArms: clusterBuildArms.concat(commanderBuildArms),